petty refactor
